load_via_python_api/requirements.txt
annotated-types==0.7.0
argcomplete==3.6.2
asn1crypto==1.5.1
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.8.0
cryptography==44.0.2
datamodel-code-generator==0.28.5
filelock==3.17.0
genson==1.3.0
idna==3.10
inflect==5.6.2
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytest==8.3.5
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
snowflake-connector-python==3.14.0
sortedcontainers==2.4.0
tomlkit==0.13.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
